samplerate needs to be a u32 not a u16

+++ src/client/backend/mod.rs
@@ -15,7 +15,7 @@ pub trait InputEndpoint {
 
 pub trait AudioSource {
 	fn next_message(&self, timeout: Duration) -> Result<Option<Message>, String>;
-	fn set_stream_info(&self, stream_info: Option<(u8, u16)>) -> Result<(), String>;
+	fn set_stream_info(&self, stream_info: Option<(u8, u32)>) -> Result<(), String>;
 	fn exit(&self);
 }
